fix(jui,example-app): remove the flawed expandToKey feature and fix "select opened file" in example app
expandToKey would require information about the whole tree, while in Tree collection we only have a subset of the tree which is currently expanded.
That makes sense because react hooks for functionality and state management of the tree are about the UI which is only concerned with that subset of the data model behind the tree.
Therefor such features don't have a place in that level.