refactor(CollectionSpeedSearch): Remove support for implicitly changing item.rendered to include highlights
Since the virtualizer renderer function is not called when speed search results are changed, we need to provide the search result matches as a context, and consume it in a component that is rendered for each item. That will ensure consistent behavior with or without virtualizer. Implicitly replacing item.rendered to a highlighted version, if it's a simple string is no longer available as a side effect of this change, but it can again be looked into.