refactor(example-app): use BrowserFS InMemory instead of LightningFS
LightningFS uses IndexedDB, which seems to have a poor performance compared to an in memory FS backend.
At least as a temporary solution, FS backend is switched to BrowserFS InMemory to have better performance.

Note that LightningFS is locally persisted, so it's mostly the first time that it's slow. But the first time matters for a demo here!