fix(List,Tree): fix a bug in fillAvailableSpace
height 100% was set which messes with the layout if the list/tree is not the only element in the parent container.
Now it's improved by using fill-available space, which also makes it work in non-flex containers too, in Chrome and Safari. Because it's not properly supported in all browsers,
in documentation of fillAvailableSpace, it's still mentioned as a requirement to have a flex container.