fix(Tree): allow horizontal overflow, which was broken since virtualization (second attempt)
In the first attempt, we tried updating layoutInfo of items to hold the actual content width, but that seemed to be problematic when the container was resized to a bigger width and then back to a smaller size.
Ended up with this new implementation where we keep the content width in a separate source (keyToWidth).
Checkout previous commit for the first implementation.