improvement(themes): remove empty light theme
The relation between intelijlaf.theme.json and Light.theme.json is not clear at the moment. We have to learn more about that. There is even no longer two light theme options in the settings, in the latest version of Intellij Community, so these two may need to be consolidated somehow.
For now it's kept in the repo, but excluded from the default *.theme.json files documented in the theming docs