refactor(actions): switch from key to code, in keyboard shortcuts
key is sensitive to Alt modifier (at least on Mac).
A mapping from keys when Alt is pressed, to the default keys doesn't seem to be straightforward, and `code` seems to be more like what shortcuts are based on. Helpers to create shortcut based on key could also be added for convenience, but the underlying data structure for shortcuts will be based on code.