feat(window): set default bounds based on content
Now `bounds` and `defaultBounds` inputs to `useResizableMovableOverlay` (and by extension, `Window`) can be `Partial<Bounds>`, which allows for more flexible default bounds. For example, one can pass `{ top, left }` to only specify the position (perhaps based on a persisted previous state), but let the sizing be still based on the content.