fix(example-app): fix some issues only happening when loaded in docs
The value of refs are updated after effects run. So the effects that use ref will work with a stale value of the ref. This only happens when the example app is loaded in docusaurus. Interestingly, the issue doesn't seem to happen outside example app, e.g. in Tooltip's documentation page.
The root cause is not clear, and rendering in StrictMode does NOT make any difference.

A workaround is used in lack of a better solution, which reruns the ref-dependant logic, asynchronously.