refactor(Popup): improve API of popup trigger and popup manager
PopupOnTrigger being an extension of Popup, wasn't a nice API, if a popup was to be implemented as a component to be used in multiple situations (e.g. with different triggers or even with PopupManager). Popup related props in could not be encapsulated within MyPopup component, which would render Popup with specific options. Similarly, PopupManager.show() expects the content of the popup, instead of the popup itself; which comes with the same shortcomings. PopupManager.show() accepting popup props not only doesn't allow for encapsulating popup-specific options in MyPopup component, but also, won't allow for stateful popup options.

Solution: There is only a couple of props that Popup(On)Trigger and PopupManager need to pass to Popup. By passing those props via Context, we can simplify the API let Popup elements be passed to both PopupTrigger and PopupManager.show().