test(example-app): experiment with performance of groupByDirectory
Trying to see how the performance of the current implementation is, and if it would be improved by voiding directory merging as post-processing, and creating the tree in one go. Turned out the performance is actually quite good in the current impl.
Keeping the other implementation tracked in git, for history.