fix(example-app): fix a crash scenario
Sometimes the blur event was not happening, so two calls to `useActivePathsProvider` from two different places would fight over the recoil state as they had different values for activePaths. One reproduction scenario was resizing commit window, while it's focused. The focus would immediately go to the editor (which itself is a bug, but unrelated), but the blur event was not triggered for some reason.
With this change, for each focused based state controller hook we keep track of the currently focused element, and update the value only the latest focused element is the same as current element. i.e. only one usage of each hook controls the state at a given time.