fix(example-app.vcs): fix issues related to refs
Two issues both related to refs not having value after the first render:
- add new file Popup is not sized correctly based on the content, and the popup doesn't get focused after open
- The first item in Search Everywhere popup is not actually focused, and pressing enter on it doesn't work, until an arrow key navigation happens.

It's not clear what the issue is exactly. It might be the same issue that's worked around by `UNSAFE_React17SuspenseFix`?

The issue only happens when the example app is loaded in the Docusaurus website, and not in Storybook.By trial and error, it became apparent that rendering `SSRProvider` is causing the issue. Perhaps it's not the root cause, but it's also not necessary to render it in ExampleContext, so it's replaced by BrowserOnly, which will render a static loading text during SSR. It shouldn't be an issue since it's not really important to render examples server-side.