feat(ModalWindow): improve imperative API by returning promise from open method
example-app is also refactored to fix  the regression in delete file(s) action. The issue started to happen after upgrading to the latest version of @react-aria/focus, but it's because the deletion happens in the confirmation dialog, and before its closed. So the node to restore the focus to is removed when the focus restoration logic runs.

Note:  upgrade of @react-aria/overlays didn't have anything to do with the fix, but was kept anyway.