refactor(collections): simplify the API for connecting an input to a collection component
useCollectionSearchInput had this advantage of allowing for connecting an input to ANY collection without requiring support from the collection component. However the collection component needed to support selectionManagerRef, so there was still a need to make the collection component compatible with useCollectionSearchInput. Plus, the usage API isn't simple enough, as it requires:
- creating a ref to hold selection manager. Passing that ref to selectionManagerRef, and importing the right type for the ref value.
- importing and invoking useCollectionSearchInput, and applying the returned props on the connected input.

That's a lot of not-so-intuitive work.
Considering any supported collection component still needs to support `CollectionRefProps` and utilize `useCollectionRef` to support connected input, the implementation is changed to have the event listeners baked into the collections (via a new useCollectionFocusProxy hook). This way at least the usage API is as simple as passing a ref to the connected input to `focusProxyRef` prop of the collection component.