test: refactor a menu test case to have it pass on CI env
For some unknown reason, scroll events were triggered upon mouse wheel, even though scroll
was prevented. It didn't happen when running locally headless with `cypress run`, so it seems like the only differentiating factor is the OS being linux in the CI env.

The test is changed to compare actual scrollY of the window, instead of checking if scroll event handlers are called.