fix tagged-docs.yaml
and also allow for triggering from the latest
state of the pipeline manually, on a specific git ref.
Useful for such cases where the tagged-docs.yaml on the desired state is broken.
Running the latest state of the action on an older state of the code is also subject to incompatibility, but having both options increases flexibility.