fix(list): change to "grid" role
After the upgrade, styled components props were renamed to have "$" prefix, so that invalid DOM props are not passed to the underlaying component.
[See more](https://styled-components.com/docs/faqs#shouldforwardprop-is-no-longer-provided-by-default).
But some tests relied on the styled list item component receiving `selected` prop.
Interestingly enough, `selected` prop wasn't even visible in the browser inspector, since it's a DOM **property** that's [not mapped](https://javascript.info/dom-attributes-and-properties#property-attribute-synchronization) to any attribute (as many properties are).

"listitem" doesn't even allow aria-selected, and the better role for List and SpeedSearchList is grid, so the roles are also changed while fixing the tests.
It's probably not quite right to omit the "gridcell" role in list rows, but a quick check with VoiceOver showed that the original Java implementation also doesn't seem to have gridcell roles.
It's just list and rows inside it.