fix(ModalWindow): fix Modal window not hiding outside content when the window content suspends
There is an useLayoutEffect call in useModalOverlay that calls `ariaHideOutside` to add aria-hidden="true" to the outside content.
If the content of the window suspended the render, the value of the ref would be null, making the useLayoutEffect noop.
That was because `useModalOverlay` was called outside the suspend boundary, but the ref was attached to an element inside the suspend boundary.